requirements.txt
greenlet==2.0.1
playwright==1.33.0
pyee==9.0.4
python-dotenv==1.0.0
typing_extensions==4.6.2
